+++ test/unit/fuzz-tools.test.js
@@ -405,6 +405,7 @@ import * as vurGradeV425 from '../../lib/vur-grade-v425.js';
 import * as gellCoombsV426 from '../../lib/gell-coombs-v426.js';
 import * as vaughanWilliamsV427 from '../../lib/vaughan-williams-v427.js';
 import * as mrcPowerV428 from '../../lib/mrc-power-v428.js';
+import * as sarnatHieV429 from '../../lib/sarnat-hie-v429.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -755,6 +756,7 @@ const MODULES = {
   'gell-coombs-v426.js': gellCoombsV426,
   'vaughan-williams-v427.js': vaughanWilliamsV427,
   'mrc-power-v428.js': mrcPowerV428,
+  'sarnat-hie-v429.js': sarnatHieV429,
   'tb-testing.js': tbTesting,
 };